mechanic change completed

+++ delete.php
@@ -11,13 +11,23 @@ $mechId = mysqli_query($conn, $sql0);
 $mechId1 = mysqli_fetch_assoc($mechId);
 $temp = $mechId1[mechName];
 
+$sql2 = "SELECT orders FROM mechanic_list where id = $temp";
+$mechId2 = mysqli_query($conn, $sql2);
+$orders = mysqli_fetch_assoc($mechId2);
+$tempOrder = $orders[orders];
+
+
 // Decresing Number of orders on delete
 $sql1 = "UPDATE mechanic_list SET orders = orders - 1 WHERE id = $temp";
 mysqli_query($conn, $sql1);
 
 // Deleting User data from Database
-$task = "DELETE FROM `appointment_list` WHERE `appointment_list`.`regNo`= '$regNo'";
-$data = mysqli_query($conn, $task);
+if ($tempOrder > 0) {
+  # code...
+  $task = "DELETE FROM `appointment_list` WHERE `appointment_list`.`regNo`= '$regNo'";
+  $data = mysqli_query($conn, $task);
+
+}
 
 if ($data) {
   # code...
